requirements.txt
Django==4.0
pre-commit==2.17.0
flake8==3.8.4
flake8-isort==4.0.0
black==22.3.0
pylint-django==2.3.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
djangorestframework-simplejwt==5.1.0
psycopg2==2.9.3
